+++ src/use-swr-infinite.ts
@@ -3,7 +3,7 @@ import { useContext, useRef, useState, useCallback } from 'react'
 
 import defaultConfig, { cache } from './config'
 import { useIsomorphicLayoutEffect } from './env'
-import SWRConfigContext from './swr-config-context'
+import SWRConfigContext from './config-context'
 import useSWR from './use-swr'
 
 import {
